refactor(bootstrap): make app more generic with kleros='>dapp-api'
